requirements.txt
langchain-core
langchain_community 
langgraph 
langchain-openai
langchain_chroma
chromadb
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
beautifulsoup4
python-dotenv
colorama
langserve
sse_starlette
uvicorn
gunicorn
fastapi
sqlalchemy
alembic
psycopg[binary]
pytest
